backend/requirements.txt
Flask
flask-cors
PyJWT
bcrypt
Flask-SQLAlchemy
flask-marshmallow
qrcode
qrcode[pil]
